requirements.txt
fastapi==0.105.0
pydantic==2.5.2
pytz==2023.3.post1
starlette==0.27.0
tortoise-orm==0.20.0
aiomysql==0.2.0
aerich==0.7.2
celery==5.3.5
vine==5.1.0
redis==3.5.3

